requirements.txt
jitsuin-archivist==0.7.0
